images: rename download_file -> download_image
