+++ server/controllers/images/lib/download_image.js
